desktop_agent/requirements.txt
pyautogui==0.9.54
pygetwindow==0.0.9
websocket-client==1.7.0
pillow==10.2.0
psutil==5.9.8
pyperclip==1.8.2
pystray==0.19.5
python-dotenv==1.0.0
